requirements.txt
amqp==5.1.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
billiard==4.1.0
celery==5.3.4
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
dnspython==2.4.2
exceptiongroup==1.1.3
fastapi==0.104.0
h11==0.14.0
idna==3.4
kombu==5.3.2
prompt-toolkit==3.0.39
pydantic==2.4.2
pydantic_core==2.10.1
pymongo==4.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
redis==4.6.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
StrEnum==0.4.15
typing_extensions==4.8.0
tzdata==2023.3
uvicorn==0.23.2
vine==5.0.0
wcwidth==0.2.8
